be more DRY by sharing functions
This is probably not the best approach as I'm including a .c file in
other .c files. But if I don't do this, I have to compile crypto_box.c
into its own unit, and if I do that, I can't make those functions (and
global variables) static. I'd like to expose nothing as this is not a
library, but a standalone C project.