remove TODO
Providing an option to explicitly create a key file shouldn't be
necessary, because if one already exists, it's only actually used when
its permissions are 0400, which excludes files that don't belong to the
user running `lock_box`. If the attacker was able to create a file as
that user, the user has other problems.